requirement.txt
azure-common
azure-mgmt-resource
azure-mgmt-storage
python-dotenv
azure-identity
azure-mgmt-sql
flask